Added babel transform runtime for polyfilling requirements
Babel transforms code it doesn't polyfill. So something like [6, -5, 8].find(x => x < 0) will not be transformed to work for IE11 for example. For this we need a polyfill. This can be done in 2 ways. Using babel-polyfill or babel-runtime. It was working before because core-js (the package that babel uses for polyfilling) was coming from other modules I have.